Revert "Merged PR 19: change price to int"
This reverts commit e9b2222e63322f782a21860fb5342ad21609fd68.

+++ Cloudents.Command/Command/UpdateUserSettingsCommand.cs
@@ -3,7 +3,7 @@
     public class UpdateUserSettingsCommand: ICommand
     {
         public UpdateUserSettingsCommand(long userId, string firstName, string lastName, string description, 
-            string bio, int? price)
+            string bio, decimal? price)
         {
             UserId = userId;
             FirstName = firstName;
@@ -18,6 +18,6 @@
         public string Description { get; }
 
         public string Bio { get; }
-        public int? Price { get; }
+        public decimal? Price { get; }
     }
 }
